+++ pkg/commands/service/domain/create.go
@@ -1,0 +1,111 @@
+package domain
+
+import (
+	"context"
+	"io"
+
+	"github.com/fastly/go-fastly/v12/fastly"
+
+	"4d63.com/optional"
+
+	"github.com/fastly/cli/pkg/argparser"
+	"github.com/fastly/cli/pkg/errors"
+	"github.com/fastly/cli/pkg/global"
+	"github.com/fastly/cli/pkg/text"
+)
+
+// CreateCommand calls the Fastly API to create domains.
+type CreateCommand struct {
+	argparser.Base
+
+	// Required.
+	serviceVersion argparser.OptionalServiceVersion
+
+	// Optional.
+	autoClone   argparser.OptionalAutoClone
+	comment     argparser.OptionalString
+	name        argparser.OptionalString
+	serviceName argparser.OptionalServiceNameID
+}
+
+// NewCreateCommand returns a usable command registered under the parent.
+func NewCreateCommand(parent argparser.Registerer, g *global.Data) *CreateCommand {
+	c := CreateCommand{
+		Base: argparser.Base{
+			Globals: g,
+		},
+	}
+	c.CmdClause = parent.Command("create", "Create a domain on a Fastly service version").Alias("add")
+
+	// Required.
+	c.RegisterFlag(argparser.StringFlagOpts{
+		Name:        argparser.FlagVersionName,
+		Description: argparser.FlagVersionDesc,
+		Dst:         &c.serviceVersion.Value,
+		Required:    true,
+	})
+
+	// Optional.
+	c.RegisterAutoCloneFlag(argparser.AutoCloneFlagOpts{
+		Action: c.autoClone.Set,
+		Dst:    &c.autoClone.Value,
+	})
+	c.CmdClause.Flag("comment", "A descriptive note").Action(c.comment.Set).StringVar(&c.comment.Value)
+	c.CmdClause.Flag("name", "Domain name").Short('n').Action(c.name.Set).StringVar(&c.name.Value)
+	c.RegisterFlag(argparser.StringFlagOpts{
+		Name:        argparser.FlagServiceIDName,
+		Description: argparser.FlagServiceIDDesc,
+		Dst:         &g.Manifest.Flag.ServiceID,
+		Short:       's',
+	})
+	c.RegisterFlag(argparser.StringFlagOpts{
+		Action:      c.serviceName.Set,
+		Name:        argparser.FlagServiceName,
+		Description: argparser.FlagServiceNameDesc,
+		Dst:         &c.serviceName.Value,
+	})
+	return &c
+}
+
+// Exec invokes the application logic for the command.
+func (c *CreateCommand) Exec(_ io.Reader, out io.Writer) error {
+	serviceID, serviceVersion, err := argparser.ServiceDetails(argparser.ServiceDetailsOpts{
+		Active:             optional.Of(false),
+		Locked:             optional.Of(false),
+		AutoCloneFlag:      c.autoClone,
+		APIClient:          c.Globals.APIClient,
+		Manifest:           *c.Globals.Manifest,
+		Out:                out,
+		ServiceNameFlag:    c.serviceName,
+		ServiceVersionFlag: c.serviceVersion,
+		VerboseMode:        c.Globals.Flags.Verbose,
+	})
+	if err != nil {
+		c.Globals.ErrLog.AddWithContext(err, map[string]any{
+			"Service ID":      serviceID,
+			"Service Version": errors.ServiceVersion(serviceVersion),
+		})
+		return err
+	}
+	input := fastly.CreateDomainInput{
+		ServiceID:      serviceID,
+		ServiceVersion: fastly.ToValue(serviceVersion.Number),
+	}
+	if c.name.WasSet {
+		input.Name = &c.name.Value
+	}
+	if c.comment.WasSet {
+		input.Comment = &c.comment.Value
+	}
+	d, err := c.Globals.APIClient.CreateDomain(context.TODO(), &input)
+	if err != nil {
+		c.Globals.ErrLog.AddWithContext(err, map[string]any{
+			"Service ID":      serviceID,
+			"Service Version": fastly.ToValue(serviceVersion.Number),
+		})
+		return err
+	}
+
+	text.Success(out, "Created domain %s (service %s version %d)", fastly.ToValue(d.Name), fastly.ToValue(d.ServiceID), fastly.ToValue(d.ServiceVersion))
+	return nil
+}